+++ sdk/attestation/attestation/recordings/node/attestationtokentests/recording_should_deserialize.js
@@ -1,0 +1,5 @@
+let nock = require('nock');
+
+module.exports.hash = "a637fc0c455eed32166fd175e0560e28";
+
+module.exports.testInfo = {"uniqueName":{},"newDate":{}}