+++ sdk/servicebus/service-bus/test/internal/unit/abortSignal.spec.ts
@@ -81,12 +81,6 @@ describe("AbortSignal", () => {
       });
 
       assert.equal(passedInOptions?.abortSignal, abortSignal);
-
-      await sender.sendMessages([testMessageThatDoesntMatter], {
-        abortSignal
-      });
-
-      assert.equal(passedInOptions?.abortSignal, abortSignal);
     });
 
     it("_trySend with an already aborted AbortSignal", async () => {